requirements.txt
allure-pytest==2.13.4
allure-python-commons==2.13.4
altgraph==0.17.4
Appium-Python-Client==3.1.1
attrs==23.2.0
beautifulsoup4==4.12.3
boto3==1.34.48
botocore==1.34.48
bs4==0.0.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
colorama==0.4.6
db-dtypes==1.2.0
distlib==0.3.8
et-xmlfile==1.1.0
filelock==3.13.1
google==3.0.0
google-api-core==2.17.1
google-auth==2.28.1
google-cloud==0.34.0
google-cloud-bigquery==3.21.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
html-testRunner==1.2.1
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
numpy==1.26.3
oauthlib==3.2.2
opencv-python==4.9.0.80
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.4.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==16.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
PySocks==1.7.1
pytest==8.0.0
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-gitlab==4.4.0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.10.0
selenium==4.16.0
setuptools==69.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2023.4
urllib3==2.0.7
virtualenv==20.25.0
wsproto==1.2.0
XlsxWriter==3.1.9
